created README template and update script
the README.md for this project will be auto-generated based on
the template included in this commit and the exceptions.yml in the
repository root.

the update.py script performs this (and any future) auto generation
from the exceptions.yml data.

the markdown is hand emitted, and could probably be written better
with a library. PRs accepted ;)